+++ src/internal/dto/lead.go
@@ -19,8 +19,8 @@ type LeadRequest struct {
 	ContactNumber  string     `binding:"required"`
 	Emailid        string     `binding:"required"`
 	Address        string     `binding:"required"`
-	CityID         uint8      `binding:"required"`
-	StateID        uint8      `binding:"required"`
+	CityID         int32      `binding:"required"`
+	StateID        int32      `binding:"required"`
 	Pincode        string     `binding:"required"`
 	CollectionType string     `json:"CollectionType" binding:"required"`
 	LeadStatusID   int8       // omitted or 0 → service uses domain.LeadStatusIDDefault (1)
@@ -56,8 +56,8 @@ type LeadUpdateRequest struct {
 	ContactNumber  *string    `json:"ContactNumber"`
 	Emailid        *string    `json:"Emailid"`
 	Address        *string    `json:"Address"`
-	CityID         *uint8     `json:"CityID"`
-	StateID        *uint8     `json:"StateID"`
+	CityID         *int32     `json:"CityID"`
+	StateID        *int32     `json:"StateID"`
 	Pincode        *string    `json:"Pincode"`
 	CollectionType *string    `json:"CollectionType"`
 	LeadStatusID   *int8      `json:"LeadStatusID"`
